pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "1.0.0"
description = "Tamplate of a discord bot"
authors = ["Sudip Roy <sudiproy20yo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {extras = ["speed"], version = "^2.3.2"}
redis = {extras = ["hiredis"], version = "^5.0.0"}
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
uvloop = "^0.17.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
pytest-asyncio = "^0.21.1"
pytest-dotenv = "^0.5.2"
pytest = "^7.4.2"
pre-commit = "^3.4.0"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--asyncio-mode=auto -v"
testpaths = ["tests"]
log_cli = "True"

[tool.semantic_release]
branch = "main"
version_variable = ["mainbot/__init__.py:__version__","pyproject.toml:tool.poetry.version"]
version_toml = "pyproject.toml:tool.poetry.version"
upload_to_pypi = false
upload_to_release = false
changelog_sections = "refactor, style, docs, breaking, feature, fix, pref"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
